requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-core==2.16.1
aio-pika==9.3.0
supabase==2.3.0
python-dotenv==1.0.0
colorlog==6.8.0

